Mocking out all web calls, to remove dependency on host site: #97. Updating Jenkinsfile to match
