+++ Jenkinsfile
@@ -1,46 +1,32 @@
-def urlBranch
-def workspace
 def branch
-def branchType
-def baseVersion
-def version
 def pullRequest
 def refspecs
 
 node {
     cleanWs()
     ansiColor('xterm') {
-        workspace = pwd()
         branch = env.BRANCH_NAME.replaceAll(/\//, "-")
-        branchType = env.BRANCH_NAME.split(/\//)[0]
-        urlBranch = env.BRANCH_NAME.replaceAll(/\//, "%252F")
-        baseVersion = "${env.BUILD_NUMBER}"
-        version = "$branch-$baseVersion"
         env.PROJECT = "selenified"
         def branchCheckout
         pullRequest = env.CHANGE_ID
         if (pullRequest) {
             branchCheckout = "pr/${pullRequest}"
             refspecs = '+refs/pull/*/head:refs/remotes/origin/pr/*'
-        }
-        else {
+        } else {
             branchCheckout = env.BRANCH_NAME
             refspecs = '+refs/heads/*:refs/remotes/origin/*'
         }
         stage('Checkout Selenified') {
             // Get the test code from GitHub repository
             checkout([
-                $class           : 'GitSCM',
-                branches: [[ name: "*/${branchCheckout}"]],
-                userRemoteConfigs: [[
-                    url          : 'https://github.com/Coveros/selenified.git',
-                    refspec      : "${refspecs}"
-                ]]
+                    $class           : 'GitSCM',
+                    branches         : [[name: "*/${branchCheckout}"]],
+                    userRemoteConfigs: [[
+                                                url    : 'https://github.com/Coveros/selenified.git',
+                                                refspec: "${refspecs}"
+                                        ]]
             ])
             sh "mkdir results"
-        }
-        stage('Update Test Site') {
-            sh 'scp public/* ec2-user@34.233.135.10:/var/www/noindex/'
         }
         try {
             stage('Run Unit Tests') {
@@ -82,11 +68,11 @@ node {
                 }
             }
             withCredentials([
-                usernamePassword(
-                    credentialsId: 'saucelabs',
-                    usernameVariable: 'sauceusername',
-                    passwordVariable: 'saucekey'
-                )
+                    usernamePassword(
+                            credentialsId: 'saucelabs',
+                            usernameVariable: 'sauceusername',
+                            passwordVariable: 'saucekey'
+                    )
             ]) {
                 stage('Execute Hub Tests') {
                     try {
@@ -103,22 +89,22 @@ node {
             }
         } finally {
             withCredentials([
-                string(
-                    credentialsId: 'sonar-token',
-                    variable: 'sonartoken'
-                ),
-                string(
-                    credentialsId: 'sonar-github',
-                    variable: 'SONAR_GITHUB_TOKEN'
-                )
+                    string(
+                            credentialsId: 'sonar-token',
+                            variable: 'sonartoken'
+                    ),
+                    string(
+                            credentialsId: 'sonar-github',
+                            variable: 'SONAR_GITHUB_TOKEN'
+                    )
             ]) {
                 stage('Perform SonarQube Analysis') {
                     def sonarCmd = "mvn clean compile sonar:sonar -Dsonar.login=${env.sonartoken} -Dsonar.junit.reportPaths='results/unit/target/surefire-reports,results/htmlunit/target/failsafe-reports,results/browserLocal/target/failsafe-reports,results/browserRemote/target/failsafe-reports' -Dsonar.jacoco.reportPaths=jacoco-ut.exec,jacoco-it.exec"
-                    if (branch == 'develop' || branchType == 'master') {
-                        sh "${sonarCmd} -Dsonar.branch=${env.BRANCH_NAME}"
+                    if (branch == 'develop' || branch == 'master') {
+                        sh "${sonarCmd} -Dsonar.branch=${branch}"
                     } else {
                         if (pullRequest) {
-                            sh "${sonarCmd} -Dsonar.analysis.mode=preview -Dsonar.branch=${env.BRANCH_NAME} -Dsonar.github.pullRequest=${pullRequest} -Dsonar.github.repository=Coveros/${env.PROJECT} -Dsonar.github.oauth=${SONAR_GITHUB_TOKEN}"
+                            sh "${sonarCmd} -Dsonar.analysis.mode=preview -Dsonar.branch=${branch} -Dsonar.github.pullRequest=${pullRequest} -Dsonar.github.repository=Coveros/${env.PROJECT} -Dsonar.github.oauth=${SONAR_GITHUB_TOKEN}"
                         } else {
                             sh "${sonarCmd} -Dsonar.analysis.mode=preview"
                         }
@@ -129,11 +115,11 @@ node {
                 jacoco()
             }
             withCredentials([
-                usernamePassword(
-                    credentialsId: 'saperstone-slack',
-                    usernameVariable: '',
-                    passwordVariable: 'botToken'
-                )
+                    usernamePassword(
+                            credentialsId: 'saperstone-slack',
+                            usernameVariable: '',
+                            passwordVariable: 'botToken'
+                    )
             ]) {
                 stage('Send Notifications') {
                     emailextrecipients([culprits(), developers(), requestor()])
