+++ src/test/java/ConflictIT.java
@@ -6,7 +6,7 @@ import org.testng.ITestContext;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-public class ConflictAIT extends Selenified {
+public class ConflictIT extends Selenified {
 
     @BeforeClass(alwaysRun = true)
     public void beforeClass(ITestContext test) {
