+++ src/test/java/integration/SelenifiedIT.java
@@ -1,22 +1,8 @@
 package integration;
 
-import com.coveros.selenified.Selenified;
-import org.testng.ITestContext;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-public class SelenifiedIT extends Selenified {
-
-    @BeforeClass(alwaysRun = true)
-    public void beforeClass(ITestContext test) {
-        // set the base URL for the tests here
-        setTestSite(this, test, "http://www.google.com/");
-        // set the author of the tests here
-        setAuthor(this, test, "Max Saperstone\n<br/>max.saperstone@coveros.com");
-        // set the version of the tests or of the software, possibly with a
-        // dynamic check
-        setVersion(this, test, "3.0.2");
-    }
+public class SelenifiedIT extends WebBase {
 
     @Test
     public void noAnnotationDetailsTest() {
